net: Extracted Port_util from TCP to net namespace
